requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.56.1
et_xmlfile==2.0.0
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.32
matplotlib==3.10.1
mlab_amo_async==0.0.5
motor==3.7.0
multidict==6.4.3
numpy==2.2.4
openai==1.75.0
openpyxl==3.1.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
prometheus_client==0.21.1
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
websockets==15.0.1
XlsxWriter==3.2.3
yarl==1.20.0
zstandard==0.23.0
aiogram==3.*
PySocks==1.7.1
psycopg2-binary>=2.9.11
rapidfuzz==3.6.2